:recycle: Moving typedefs
